+++ bazaarapp/strings.py
@@ -55,40 +55,6 @@ stages_info = [
             ],
         ],
     },
-    # {  # 6
-    #     "text": "💰 Выберите ценовую категорию продажи",
-    #     "text_buttons": [
-    #         [
-    #             {
-    #                 "text": "до $1000",
-    #             },
-    #             {
-    #                 "text": "от $1001 до $2500",
-    #             },
-    #         ],
-    #         [
-    #             {
-    #                 "text": "от $2501 до $5000",
-    #             },
-    #             {
-    #                 "text": "от $5001 до $7500",
-    #             },
-    #         ],
-    #         [
-    #             {
-    #                 "text": "от $7501 до $10000",
-    #             },
-    #             {
-    #                 "text": "более $10000",
-    #             },
-    #         ],
-    #         [
-    #             {
-    #                 "text": "Отменить",
-    #             },
-    #         ],
-    #     ],
-    # },
     {  # 7
         "text": "💸 Укажите цену в долларах\n\n",
         "text_buttons": [
