+++ garage_band_bot/settings.py
@@ -155,6 +155,9 @@ BAZAAR_FILTER_LIVE_TOKEN = os.getenv("BAZAAR_LIVE_TOKEN")
 BAZAAR_PUBLISH_ID = os.getenv("BAZAAR_PUBLISH_ID")
 BAZAAR_TEST_PUBLISH_ID = os.getenv("BAZAAR_TEST_PUBLISH_ID")
 
+REPAIRS_FILTER_LIVE_TOKEN = os.getenv("REPAIRS_FILTER_LIVE_TOKEN")
+REPAIRS_FILTER_TEST_TOKEN = os.getenv("REPAIRS_FILTER_TEST_TOKEN")
+
 DEV_TG_ID = os.getenv("DEV_TG_ID")
 
 # media
